refactor files and minor tweaks
